cpp09

+++ cpp09/ex00/BitcoinExchange.hpp
@@ -5,12 +5,13 @@
 # include <stdexcept>
 # include <fstream>
 # include <ctime>
+# include <cstdlib>
 #include <map>
 
 struct pair
 {
     std::string key;
-    float value;
+    double value;
     int valide;
 };
